requirements.txt
dialogflow==0.7.2
python-telegram-bot==12.2.0
requests==2.22.0
